添加安全验证

+++ src/main/java/com/xinshijie/gallery/domain/Image.java
@@ -26,7 +26,7 @@ public class Image {
 
     private String url;
 
-    private Long aid;
+    private Integer aid;
 
     private String sourceUrl;
